fix: use only float instead of double and fix parenthesis order

+++ include/cuda_path_tracer/vec3.cuh
@@ -66,4 +66,4 @@ __device__ auto roundScatterDirection(const Vec3 &direction,
                                       const Vec3 &normal) -> Vec3;
 __device__ auto reflect(const Vec3 &v, const Vec3 &n) -> Vec3;
 __device__ auto refract(const Vec3 &v, const Vec3 &n,
-                        double eta_component) -> Vec3;
+                        float eta_component) -> Vec3;
